Checking in Aniruddha's producer/consumer taskpool implementation of the Fock matrix computation on which the HIPS paper is based.  I didn't realize I hadn't done this before.  The files in question are:         fock-3.chpl         taskpool.chpl
Unfortunately, this test stopped working in an
undetected manner between v0.5 and v0.6 due to
the use of an array of sync variables within
a class (captured in a future in my separate
checkin earlier today).  For this reason, I'm
checking this in as a .notest for now.

Checking in my rewrite of fock-3.chpl which is
made up of the files:

     fock-3-blc.chpl
     taskpool-blc.chpl
     blockindices-blc.chpl (I broke the blockIndices
                            class into its own file)

I implemented a workaround for the sync arr in
class problem by making a global array of sync
variables (of fixed size which is presumed to be
bigger than poolSize, but not checked).  This
implementation also includes a number of refactorings
and renamings aimed at making the code clearer and
cleaner.  Some additional modifications that I have
in mind are noted in the comments.

This code also matches the excerpts-blc.chpl file
in the hips08 subdirectory, as validation that it
can be made to work (modulo the future mentioned
above).

Some of the main modifications from fock-3.chpl are:
- moving the very complicated loop in buildjk_atom4()
  into an iterator on the blockIndices class
- moving declarations of values used to drive this
  complicated loop into the loop itself and simplifying
  their initialization somewhat
- using a nil class pointer as the sentinel blockIndices
  value rather than a block whose indices are 0's
- collapsing 2 loops in a 4-deep loop nest into 1 using
  a tensor iteration
- making the default initialization for numConsumers
  be based on the number of cores available across
  the Locales array minus 1 (or 1 if this value is 0).
- renaming numLocs to numConsumers and loc to cons
- moving the declarations of bI and copyofbI to their
  definition points
- using $ in sync var identifiers
- changing the blockIndices class to store ranges
  rather than scalars

Primary remaining pieces of work are:
- make the bas_info array into an array of ranges
- add the use of array views for the variables
  dij, dik, dil, etc.
- add support of arrays of arrays of different
  sizes and use this to replace groups of six
  explicit statements in buildjk_atom4() into
  a loop over a 6D array


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13155 3a8e244f-b0f2-452b-bcba-4c88e055c3ca