Make the first field in derived classes be "super" - the value type of the class they inherit from. Previously, the fields defined in the superclass were directly injected into the subclass.
Until codegen, the super field is treated as a regular class object with the
type of the super class. In codegen, an extra underscore is added to it's type
name in order to generate it as a record. Also, add the reference operator
(&) whenever it is accessed. Since the compiler thinks super is a regular
class, not a value, accesses to it need this special treatment to make them
more class-like.

The default constructor for a class takes the same arguments as the default
constructor for its super class, plus one for each field defined directly by
the class. All arguments to the default constructor can have default values
which are taken from the field declaration point.

Added an invisible-to-the-user argument as the last argument to default
constructors. It's type is the same as the type of the class to construct.
For calls into the constructor that should allocate a new class it has a nil
default value. Once a class is allocated and the _cid field is set, the super
class default constructor is called with it's arguments filled in and the
extra argument set to "this.super" (which has the same memory address as
"this"). The super class can then set up its fields and not allocate any
memory since it was already allocated by the subclass. The extra argument is
currently named "meme". This way it will be pretty easy to do a search and
replace on it once I think of a decent name. =)

Update the code for creating writer methods for classes. Instead of printing
just fields in the class, the super class fields are recursively printed
followed by the locally defined class fields. A similar change needs to be
done for the reader methods.

Added the "no object" pragma to some compiler generated class symbols created
for parallel constructs and iterators. In the future, if a class doesn't
extend object, it should have this pragma.

Update the ClassType::getField method to go up the inheritance hierarchy
looking for a field instead of only looking through the fields directly defined
in the current class. Also, previously it was an internal fatal error to
call getField on a name not defined by the class. Now it returns NULL in that
case. It might be good to add an argument to this method to decide how severe
not finding the named field is that defaults to fatal error, but I haven't
done that.

Update the SymScope::lookup and SymScope::lookupLocal methods to treat all
superclass fields as nearer in scope than any other outer scope.

Codegen classes breadth first starting with "object", then all of its dispatch
children, then theirs, etc. This is required so the super field's type is
completely defined before it is used.

Paired with Steve: added a runtime macro _WIDE_CLASS_GET_SUPER that gets the
super field as a wide reference. Also got some help fixing a few C compiler
warnings when compiling with --no-inline.

Removed four .futures about shadowing fields that now work. I didn't do a full
.futures test run, just tried the most obvious places, so with luck, more may
show up tonight.

Changed two tests that attempt to inherit from generic classes from passing
tests into semantic futures. They attempt to instantiate the generic fields
from the derived class constructor. It should probably be instantiated at the
declaration of the subclass - after the colon.

Expect garbage collection test classes/diten/allocateManyObjects to fail.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13182 3a8e244f-b0f2-452b-bcba-4c88e055c3ca