Changed existing internal compiler flag --ccwarnings to --[no-]cc-warnings and made its default value follow the setting of CHPL_DEVELOPER/--devel -- warnings are on for developers, off for non-developers.  This flag allows that default to be overridden and turns on/off warnings for the generated code.
Changed the compiler-oriented Makefiles to define a new
variable WARN_GEN_CFLAGS that will be used to define
the flags to be used on the generated code when
--cc-warnings is turned on.  This replaces the old scheme
in which these warnings were explicitly turned on in
the Makefiles themselves.  Added a comment noting the
new behavior to avoid replacing this logic in the future.

Added what I hope will be appropriate WARN_GEN_CFLAGS
settings for Intel, Pathscale, and PGI compilers.  None
of our Cray compilers have ever had warning flags specified
for developers and I did not add them with this check-in.

Made start_test set -compopts equal to "--cc-warnings" by
default such that all executions of the test system,
whether from nightly cron jobs or by hand, will get warnings
on the generated code.  Also made the -compopts option
additive so that new compiler options can be added to the
existing set (rather than using quote-delimited lists of
compiler options).  In this way, you can turn off warnings
on the generated code for a test run if you want to using
-compopts "--no-cc-warnings".

Have tested this on 64-bit linux and it looks clean.
Today was a quiet day for checkins so it seemed like
a reasonable time to check it in and see how it does
with the other compilers and platforms.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13208 3a8e244f-b0f2-452b-bcba-4c88e055c3ca