Moved the Xcode project from the top level to the etc directory and adjusted all the paths in the project file accordingly.
Removed chapel.tab.c and lex.yy.c from the project file, and added shell scripts
to create these files (though in a directory more related to the one containing
.o files).  (These .c files need to be compiled by the C++ compiler, even though
they actually contain C code, so the normal rules for processing yacc and lex
source files can't be used.)

Added COPYRIGHT and LICENSE to the project, and added a shell script to process
these files so they can be #include'd as required.  (Again, the processed files
are created in a directory more related to the one containing .o files.)

Added more warning flags, though warnings regarding comparisons between signed
and unsigned expressions had to be disabled, since the output of flex contains
such a comparison.

With these changes, it should be possible to do a fresh check out from scratch
and build the entire compiler and runtime using Xcode.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13220 3a8e244f-b0f2-452b-bcba-4c88e055c3ca