Added semantic support for the 'new' keyword.  Use 'new' to construct a class; elide 'new' to specify a class type.  This is work that I started with David the afternoon of 1/25 when he was in town and that has monopolized my time since.  This is a large change.
Added a second default constructor for every class that is used when
'new' is elided.  This is the type constructor and is captured in
Type::defaultTypeConstructor.  This function returns a type and takes
as arguments only the generic fields of a class or record.

Added Symbol* FnSymbol::_outer which is set for nested class
constructors and replaces the use of _this in the constructor.  Nested
class constructors now take an argument called outer and still use
_this to refer to the constructed inner class.

Reworked destructuring of tuples and parsing of tuples to rely on the
_build_tuple function which takes a variable number of arguments
rather than the tuple constructor.

Changed the _init function into a primitive called PRIMITIVE_INIT.
Array types, which are translated into runtime values, still use a
_this function.  In order to support records and tuples of arrays, I
believe we will need to switch back to at least using an _init
function (and doing so correctly, i.e., having _init take a single
type formal argument rather than a value) since a record that contains
an array type will need to store value information at runtime too.
This has never really been supported so I futurized the test
arrays/deitz/test_tuple_of_array.chpl which was certainly not working
correctly.

Added PRIMITIVE_NEW which is removed in normalize and marks uses of
the 'new' keyword in Chapel codes.

Removed pragma "constructor type" which is no longer necessary.

Inserted checking to make sure that types are not assigned to values
and vice versa.  Tightened up some compiler-generated code that took
advantage of this looseness.

Applied much special-casing of build_sparse_subdomain_type to keep
sparse domains working even though they are not generally supported
due to the lack of runtime sparse domain type information.

Renamed function resolutions setFieldTypes to more accurately reflect
what it is doing: setScalarPromotionType and fixTypeNames.

Improved the check on fields with generic types to result in cleaner
error messages.  This is important because ignoring this catch
resulted in a much more difficult to diagnose situation later.

Added support for variable number of arguments of intent type in order
to support a _build_tuple type function.  This should allow tuples and
tuple types to be parsed identically.

Added code during function resolution to resolve the type constructor
when attempting to resolve a constructor so that the type constructor
is always resolved first.  An out-of-order function resolution engine
would avoid this need.  Also added special handling of an array's
_init function for similar reasons.

Updated code in the standard modules to use 'new' when constructing a
class.

Added several tests of errors resulting from omitting 'new'
incorrectly.

Removed a future for a test of a type method with a default value.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13264 3a8e244f-b0f2-452b-bcba-4c88e055c3ca