Fixed a bug in which references to unions could be lost.
Added support for setting and getting members of unions via wide
references.  Added support for setting and getting the union id of
unions via wide references.

Changed the primitive that gets the id of a union from checking the id
and returning bool to returning the id itself.

This fixes regressions seen under gasnet.  The following issues are
expected:

  arrays/deitz/part2/test_array_init
  arrays/deitz/test_array_dot_type
  classes/bradc/overloadMethods/v2NoReturnValueButType
  classes/deitz/tree
  execflags/bradc/gdbddash/declint
  execflags/bradc/gdbddash/declint2
  execflags/bradc/gdbddash/gdbSetConfig
  functions/deitz/iterators/tree
  memory/shannon/outofmemory/mallocOutOfMemory
  parallel/cobegin/bradc/varsEscape-workaround
  release/examples/iterators
  studies/uts/Makefile
  trivial/deitz/test_type_sizes

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13280 3a8e244f-b0f2-452b-bcba-4c88e055c3ca