Fixed a bug in expanding tuples using (...t) syntax for the case where t is a tuple type.
Expanded tuple types for the implementations of the read and readln
functions.

Moved the implementations of read and readln methods and functions
that take multiple types into the IO module.  Such functions return a
tuple of values.

Updated the test case to use the code in the standard modules and test
all four functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13300 3a8e244f-b0f2-452b-bcba-4c88e055c3ca