Make strings into wide strings. This required special casing string literals being passed to functions and primitives by putting them into temps at the call site.
Wide strings have an extra size field that other wide classes don't have to
hold the string length.

Also, added macros _COMM_WIDE_GET_STRING and _SET_WIDE_STRING to interface
strings with the comm-layer. No PUT_STRING like macro was needed because
all string puts are done by building a local wide string and putting that
to the remote location.

Changed the raddr argument to _chpl_comm_get from a "void*" to a "const void*"
because it's more descriptive and otherwise passing a Chapel string to it
drops the constness.

Fixed a typo in checking if wide classes are not equal to each other - changed
&& to ||, so now they're not equal if addr1 != addr2 || locale1 != locale2.

Added a bunch of tests accessing and setting remote strings.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13316 3a8e244f-b0f2-452b-bcba-4c88e055c3ca