Added an initial capability for skipping over certain test directories based on environment variable settings.  This will continue to improve over the next few days until we're a little happier with it.
The current support allows a test directory to
contain a SKIPIF file which looks approximately
like the following:

     -- SKIPIF -------------------------------------------
     # skip this directory for gasnet or non-Mac platforms

     CHPL_COMM == gasnet
     CHPL_TARGET_PLATFORM != darwin
     -----------------------------------------------------

In particular, each line can either be a comment, a
blank line, or contain an equality or inequality
comparing an environment variable with an expected
value.  All of the environment variable settings will
be evaluated and or'd together -- if any of them are
true, the directory will be skipped.

Here are some ways this needs to evolve:

- probably need to support some sort of logical-and
- need to be able to test for things like --compopts
  contains "--baseline" as one of its options
- probably want to support these on a file-by-file basis


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13320 3a8e244f-b0f2-452b-bcba-4c88e055c3ca