Switched the mechanism to spawn local threads in gasnet to use _chpl_begin rather than a direct call to pthread_create via the chpl_thread_create macro.  Removed said macro.  With Samuel's change earlier to remove the necessity _chpl_cobegin, I believe _chpl_begin is the only place where we spawn threads.
Removed _chpl_thread_t as it is now specific to threads-pthreads.

Set the config const maxThreads to the special value 3 before calling
into the code to initialize module since this will have to use
_chpl_begin to spawn threads.  Otherwise it could be zero or
uninitialized resulting in a problem within this function.  It is set
to the real number when the config const is set in the initialization
function of the ChapelBase module.

Updated stale code in the PRINTF debug macro for comm-gasnet.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca