Adding a.out and output to $CHPL_HOME ignores ---------------------------------------------
This is primarily a test of the new SVN Commit log
message format, and a trial of a concept that Sam
and I independently came up with -- including the
first line of the log message as the subject line
of the mail message.  Should make email conversations
leading from check-in messages easier to read.
See what you think.

Under CVS, we had the ability for everyone to
create their own local .cvsignore file in the
root directory containing .cvsignore and anything
else they wanted to ignore.  I haven't been able
to find an equivalent in subversion yet, so am
adding a.out and output to the svn:ignore properties
of the root directory, under the assumption that
everyone else probably creates these as often as I
do.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca