Add a test that goes remote in a begin.
This test does an on in a begin.  The main thread spin waits for it to change
a variable.

Also, added a NUMLOCALES file and removed the .numlocales files that had the
same value as the default. Modified the sub_test script to use the default
value if a .numlocales doesn't exist. This now works just like the other
settings that can be set both globally and with a .option file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13434 3a8e244f-b0f2-452b-bcba-4c88e055c3ca