fixes uninitialized vars in iterator methods (extends r13417)
Added code to the getNextCursor and zipCursor1 iterator methods
(previous checkin just handled getHeadCursor) to initialize variables
of record type with what is already in the iterator class.  This is
necessary because we write these variables back to the iterator class
at the end of these iterator methods.  This is only done for records
because a write to a record can be indirect (via a reference) and so
it is more difficult to determine where the record writes exist.  I'm
not sure this additional code in these other two methods is necessary,
but it seems like it should be even if the problem is not showing up
in any current test cases.

Added code to the iterator (after it is changed into a function that
returns an iterator class) to initialize all fields that are records
to avoid C warnings about possibly uninitialized variables.  This
should avoid a warning when compiling with -O3 and fix the RA
regression in the opts testing.  I was wrong in r13417 because I never
looked at the code after turning inlining back on.  This change also
fixes the concern I had that the field in the iterator class is
uninitialized because now it is.

Inserted a nil temp for the PRIMITIVE_SET_MEMBER case.  This started
showing up because of the initializations to default values inserted
in this change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13437 3a8e244f-b0f2-452b-bcba-4c88e055c3ca