multilocale support for sync vars on locale 0
Added support for sync variables on locale 0 by inserting an
on-statement into the functions that lock and unlock sync variables
via primitives.  The primitives then assume that the sync variable is
local and call into pthread functions.

Optimized _chpl_comm_fork to not fork a process but rather just
evaluate the function if execution is already occuring on the "remote"
locale.  This optimizes on-statements onto the current locale.

Generalized disabling of heap accesses on arguments within where
clauses to be anywhere other than within the function's body.

Added code to handle insertion of line numbers into functions that are
called indirectly via _chpl_comm_fork or _chpl_begin.

Added sync primitives to the list of primitives that dereference their
actuals.

Removed a future.

Note that remote writes are still not working.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13460 3a8e244f-b0f2-452b-bcba-4c88e055c3ca