Capture of Aniruddha madness bug
Checking in a copy of Aniruddha's madness bug that I've been
working on for the past few days -- it is more or less the
same code that he currently has checked in with a few
differences -- the class is changed to a record (required
to reproduce the bug); I put in some debug print loops;
and I put in a loop that shows the error before it happens
in the code without it.  I also took out the extraneous
files and made the test not be Makefile-based.

See the .future for futher notes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13462 3a8e244f-b0f2-452b-bcba-4c88e055c3ca