Fixed .length method for degenerate ranges
The .length method on degenerate ranges would sometimes
return a negative number.  In particular 1..-1 would
say that it had a length of -1 rather than 0.  Fixed
this which also fixes the degenerate domain bug that
I filed some time ago (in which memory was allocated
for degenerate domains -- a lot of memory for that
matter).

Thought that this was part of the problem with the
Aniruddha bug that I just filed, but it turned out
not to be.  But the fix was easy and good, so might
as well check it in.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13464 3a8e244f-b0f2-452b-bcba-4c88e055c3ca