Removing unnecessary NOEXEC file.
This NOEXEC file is no longer necessary, and also
causes the executable that's created to not be
deleted, so I'm removing it.  (The more general
fix would be to have the sub_test script remove
the binary even when NOEXEC is true, but this was
more delicate than I was willing to undertake
right now).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13466 3a8e244f-b0f2-452b-bcba-4c88e055c3ca