Fixed pretty-printing bug with --cpp-lines
Fixed a previously uncaptured bug in the compiler
in which long lines in the generated C code would
be broken up by #line directives when the --cpp-lines
flag was used, due to the use of a fixed-size buffer
to process those lines during the beautification
pass (which handles pretty-printing and indentation
of generated code).  Filing a test against the
problem and the fix.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13467 3a8e244f-b0f2-452b-bcba-4c88e055c3ca