Replacing CVSLOG with SVNLOG
Since SVN stores CVSLOG-like information (and since
the mechanisms we used to update CVSLOG don't make
sense for SVN), removed CVSLOG and added a script
called updateSVNLOG.pl that will create an SVNLOG
file for you that is up-to-date with your copy of
$CHPL_HOME (Note that CHPL_HOME must be set in order
to use this script).

The first time you run this, it will take a little
while, due to the time required to list the entire
set of revisions.  On subsequent runs, only the
incremental portion of the log will be generated
and concatenated onto the end of SVNLOG.  A hidden
file -- .SVNLOG-revnum -- is used to track which
revision your log currently reflects, and is used
by the script to make incremental updates.

To get a fresh SVNLOG remove the SVNLOG file and/or
the .SVNLOG-revnum file and re-run the script.

Please let me know if you have any problems with
this or feature requests (note that the format
is the default format used by svn, so more work
would be required to get something a little
prettier.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13469 3a8e244f-b0f2-452b-bcba-4c88e055c3ca