added polling function to ensure gasnet progress
Added a polling function that is spawned in a separate thread when
GASNet is used.  This polling function calls into GASNET_BLOCKUNTIL in
order to ensure that remote puts and gets are handled even if any
computation on the remote locale is purely local.

Unfortunately this uses pthread routines directly in the gasnet comm
code rather than code in our threads runtime.  I'll rework this by
adding another function to the threads runtimes that allow threads to
be spawned before the threads runtime is initialized since we want to
spawn the thread when we initialize the comm runtime which happens
before we initialize the threads runtime.

Uncommented code to free the data that is sent to the remote forking
function.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13472 3a8e244f-b0f2-452b-bcba-4c88e055c3ca