improved robustness of .locale
Made .locale work on classes and fields accessed within classes.

Removed special-casing code in normalization that short-circuited
inserting a heap access (for locals and globals put on the heap) if
the access was in the context of ".locale".  This worked for certain
cases but not in general.  In essence, it made us take the location of
the heap class rather than what was in the heap class which could
itself be a wide class.

Added special-casing code to cullOverReferences so that we do not
dereference a reference if the reference is accessed in the context of
".locale".  This is a more general way to handle the issue above and
allows us to determine the locale of a member access since the wide
reference is not dereferenced.

Added runtime macro _COMM_WIDE_GET_LOCALE to handle the case where we
want to determine the locale of a wide reference to a wide class.
This requires communication to determine the locale that the wide
class is allocated on.

Alternatively we could return the locale that the reference exists on
instead of the wide class.  I think the common case is to want to know
where the class is allocated rather than where the reference lives
especially given that class references are copied.  For example, "on
this" would always yield the current locale since 'this' is copied.  I
think we should keep this issue in mind moving forward.

Added two working test cases.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13474 3a8e244f-b0f2-452b-bcba-4c88e055c3ca