Make the statement "a;" be a read of "a"
Add an inline function called _statementLevelSymbol that takes a single
argument and immediately returns it to ChapelBase. Also add a param version.

SymExprs that are equal to their getStmtExpr() are statement level symbol
accesses. Find them and pass them through a call to the function added above.

Then, accessing a sync variable as a statement translates into a readFE and
accessing a single variable translates into a readFF.

Other statement level accesses are inlined into no-ops.

Update Brad's simplereadSync test so it enforces this behavior. Previously,
it didn't require the sync to actually be read when printing that it was.

Added a similar test for a single var.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13476 3a8e244f-b0f2-452b-bcba-4c88e055c3ca