support for remote sync vars
Added support for remote sync and single variables.  These should now
work on any locale, not just locale 0.

Changed all occurrences of "on Locales(0)" in the base module to "on
this" instead.

Removed a line that disabled wide versions of sync and single classes
from being created.

Added code to handle primitives on wide sync and single classes that
assumes they are local (which they are thanks to the "on this" line).

Note that writeln as well as all other file functionality does not
work on remote locales because remote files are not yet working.
These should be relatively easy to enable using the same "on this"
technique.  The main issue for writing to stdout is that the C file
"stdout" is different in different processes.

Removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13478 3a8e244f-b0f2-452b-bcba-4c88e055c3ca