added "end" functionality
Added support for an end-statement that waits for all dynamically
encountered begin-statements to terminate.

Thus

  end {
    begin stmt1();
    begin stmt2();
  }

is similar to

  cobegin {
    stmt1();
    stmt2();
  }

but there are some important differences.  First, if begin-statements
are dynamically encountered when stmt1() or stmt2() are executed, then
the former code will wait for these begin-statements to complete
whereas the latter code will not.  Second, although this is not
currently the case, I think we want to implement cobegin in such a way
that the top-level threads are started before any dynamically
encountered threads are spawned even if we don't make any guarantees.

Changed uses of "end" as a variable name to something else, usually
"stop."

Added semantics of wrapping main in this end-statement.  This handles
termination so that we do not exit until all of the threads are
finished.  To exit early, the programmer must now make an explicit
call to exit().

This implementation work does not impact the runtime.  It is built on
a new class added to the base module called _EndCount that contains a
sync int and a sync bool.  The sync int is used to keep track of the
number of dynamically encountered begin-statements.  The sync bool is
used to signal the thread that executed the end statement that it can
continue.  Functions _upEndCount() and _downEndCount() increase and
decrease the count on the current count.  This cound is accessed via
primitives that are removed during the parallel pass and replaced by
threading the count through the program as a formal argument where
necessary.  This works like insertLineNumbers does with line numbers
and filenames.

Note that this construct and the functions that work on specific
instances of the _EndCount class can be used to rework cobegin and
coforall to avoid the stack of sync variables.

Added end to the list of keywords in the spec and updated the vim and
emacs mode files.  Added initial definition of end to the spec.

Removed three multilocale termination futures.

Added three tests of this new statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13479 3a8e244f-b0f2-452b-bcba-4c88e055c3ca