fix COMM=none bug due to 'collision'
Fixed a bug caused by my change from "on Locale(0)" to "on this" in
getting remote sync variables to use when taken with David's change to
get statement-level sync variable accesses to work.  With --local
which is the default when COMM=none is used, the expression in the
on-statement is evaluated only for side effects.  But then evaluating
"this" only for side effects from within the readFE method resulted in
a call to the readFE method.

I made it so that using the --local flag does not evaluate the
on-expression at all.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13480 3a8e244f-b0f2-452b-bcba-4c88e055c3ca