initialization of maxThreads + termination of gasnet polling thread
Changed the way maxThreads is initialized if the user does not specify
a value.  It is determined by two functions, one in the comm layer and
one in the threads layer.  The result is taken to be 0 (unlimited) if
they both specify this.  Otherwise the result is taken to be the
minimum number specified.  The current defaults are specified as
follows: threads-mta specifies _coresPerLocale() * 100; comm-gasnet
specifies 255 (256 limit less one for the polling thread); all other
threads and comm layers leave it unlimited.

Changed the gasnet polling function to BLOCKUNTIL the per-process
integral value exitSignal is one.  It is initialized to zero and set
to one using a local gasnet_put in the _chpl_comm_exit routines.  This
effectively kills the polling thread.  I think this is desirable
because otherwise that thread is running gasnet routines even as we
try to exit gasnet.

This increases gasnet stability but I'm still seeing spurious errors
probably related to termination.  Just not as many.  For example, I
saw

    *** AMUDP ERROR: I just refused a message and returned to
    sender. Reason: EBADTAG

on one test (test/portability/compIntSizes) but could not repeat it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13487 3a8e244f-b0f2-452b-bcba-4c88e055c3ca