Support for .skipif / Futures always listed in test summary
Added a <testname>.skipif capability similar to the directory-level
SKIPIF files.  A file-level .skipif will be ignored if the test
is explicitly named on the start_test line (much as .future and
.notest tests are).  This seems to work for the .skipif file that
Steve checked in this morning.

Also, previously our test summary would only print #Futures in
the summary information if the invocation of the start_test
script requested futures to be tested.  In this check-in I've
changed this to always print the number of futures since futures
may now be run by naming them explicitly on the start_test
command.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13501 3a8e244f-b0f2-452b-bcba-4c88e055c3ca