support for multilocale write and i/o
Refactored ChapelIO module somewhat and inserted "on this" in front of
file methods that need it.  That is, writing to a file is now always
done where that file class exists.  This is essential even for writing
to stdout since the C file stdout has a different address in different
processes.

Removed a future of writing an array on a remote locale and added a
simple test of writeln.

Note that _debugWriteln never requires launching a thread on another
locale nor does it do any synchronization.  For these reasons, I think
it is worthwhile exposing this to the user somehow, though I don't
know what we should call it or whether it should support full write
functionality.  Currently it only prints strings or anything that can
be coerced to a string.  Writing to a string is a work-around.  Note
that writing to a string is not locked either and also does not
require launching a thread on another locale.  Should writing to a
string be locked?  My thinking is yes, though it is currently not.

Avoided a bug in which an argument of out/inout intent is updated in
an on-statement (assuming a simple insertion of an on-statement into
the read function).  Added a future to capture this issue.

Added a test of reading from a remote file.

Added a future of remotely reading a class.  This fails because the
default function built for reading a class does not migrate on the
locale that the file class exists on before calling readLitChar.
Looking into a fix now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13504 3a8e244f-b0f2-452b-bcba-4c88e055c3ca