Made send_/recv_patch use ~/patches directory
This change makes our send_patch and recv_patch
scripts use the ~/patches directory (and send_patch
will create it if it doesn't exist) rather than ~/
simply to avoid cluttering up home directories.

This change does not include either of the other
proposed changes to these scripts that we've been
discussing (due to lack of clarity about which to
implement).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13507 3a8e244f-b0f2-452b-bcba-4c88e055c3ca