improved support for extern functions on multilocale
This checkin should fix the four UTS regressions on gasnet.

Added codegen support for assigning a wide reference to a reference.
This is done for extern functions that expect references (via inout or
out intents) since we assume the wide references is local.  There is
no check.  Support for extern functions is very limited.

Fixed a thinko in codegening calls to extern functions in which the
formal was checked to see whether & needs to precede the actual rather
than the actual itself.

Made it so that temps inserted to handle removing the wide part of a
reference or class for a call to an extern function are not
initialized by dereferencing the wide reference if the temp is itself
a reference.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13513 3a8e244f-b0f2-452b-bcba-4c88e055c3ca